Make AdHocMapper and RelationalAdHocMapper public
Part of #32680
