+++ src/EFCore/Metadata/AdHocMapperDependencies.cs
@@ -1,0 +1,55 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace Microsoft.EntityFrameworkCore.Metadata;
+
+/// <summary>
+///     <para>
+///         Service dependencies parameter class for <see cref="AdHocMapper" />
+///     </para>
+///     <para>
+///         This type is typically used by database providers (and other extensions). It is generally
+///         not used in application code.
+///     </para>
+/// </summary>
+/// <remarks>
+///     <para>
+///         This is an internal API that supports the Entity Framework Core infrastructure and not subject to
+///         the same compatibility standards as public APIs. It may be changed or removed without notice in
+///         any release. You should only use it directly in your code with extreme caution and knowing that
+///         doing so can result in application failures when updating to a new Entity Framework Core release.
+///     </para>
+///     <para>
+///         The service lifetime is <see cref="ServiceLifetime.Scoped" />. This means that each
+///         <see cref="DbContext" /> instance will use its own instance of this service.
+///         The implementation may depend on other services registered with any lifetime.
+///         The implementation does not need to be thread-safe.
+///     </para>
+/// </remarks>
+public sealed record AdHocMapperDependencies
+{
+    /// <summary>
+    ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
+    ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
+    ///     any release. You should only use it directly in your code with extreme caution and knowing that
+    ///     doing so can result in application failures when updating to a new Entity Framework Core release.
+    /// </summary>
+    [EntityFrameworkInternal]
+    public AdHocMapperDependencies(
+        IModel model,
+        ModelCreationDependencies modelCreationDependencies)
+    {
+        Model = model;
+        ModelCreationDependencies = modelCreationDependencies;
+    }
+
+    /// <summary>
+    ///     The model source.
+    /// </summary>
+    public IModel Model { get; init; }
+
+    /// <summary>
+    ///     The convention set to use when creating the model.
+    /// </summary>
+    public ModelCreationDependencies ModelCreationDependencies { get; init; }
+}
