+++ src/EFCore/Metadata/IAdHocMapper.cs
@@ -6,6 +6,18 @@ namespace Microsoft.EntityFrameworkCore.Metadata;
 /// <summary>
 ///     Creates ad-hoc mappings of CLR types to entity types after the model has been built.
 /// </summary>
+/// <remarks>
+///     <para>
+///         The service lifetime is <see cref="ServiceLifetime.Scoped" />. This means that each
+///         <see cref="DbContext" /> instance will use its own instance of this service.
+///         The implementation may depend on other services registered with any lifetime.
+///         The implementation does not need to be thread-safe.
+///     </para>
+///     <para>
+///         See <see href="https://aka.ms/efcore-docs-providers">Implementation of database providers and extensions</see>
+///         for more information and examples.
+///     </para>
+/// </remarks>
 public interface IAdHocMapper
 {
     /// <summary>
